requirements.txt
altair==4.1.0
argcomplete==1.12.3
argon2-cffi==21.1.0
astor==0.8.1
astroid==2.7.3
attrs==21.2.0
backcall==0.2.0
backports.zoneinfo==0.2.1
base58==2.1.0
beautifulsoup4==4.10.0
billboard.py==6.2.1
bleach==4.1.0
blinker==1.4
bokeh==2.2.0
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==7.1.2
cycler==0.10.0
debugpy==1.4.1
decorator==5.0.9
defusedxml==0.7.1
entrypoints==0.3
filelock==3.0.12
gitdb==4.0.7
GitPython==3.1.18
huggingface-hub==0.0.16
idna==3.2
importlib-metadata==4.8.1
ipykernel==6.3.1
ipython==7.27.0
ipython-genutils==0.2.0
ipywidgets==7.6.4
isort==5.9.3
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
jsonschema==3.2.0
jupyter-client==7.0.2
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.1
kiwisolver==1.3.2
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.2
mccabe==0.6.1
mistune==0.8.4
nbclient==0.5.4
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
nltk==3.6.3
notebook==6.4.3
numpy==1.21.2
packaging==21.0
pandas==1.3.2
pandocfilters==1.4.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.2
platformdirs==2.3.0
prometheus-client==0.11.0
prompt-toolkit==3.0.20
protobuf==3.17.3
ptyprocess==0.7.0
pyarrow==5.0.0
pycparser==2.20
pydeck==0.7.0
Pygments==2.10.0
pylint==2.10.2
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
pyzmq==22.2.1
regex==2021.8.28
requests==2.26.0
sacremoses==0.0.45
scikit-learn==1.0
scipy==1.7.1
seaborn==0.11.2
Send2Trash==1.8.0
sentence-transformers==2.0.0
sentencepiece==0.1.96
six==1.16.0
smmap==4.0.0
soupsieve==2.2.1
streamlit==0.88.0
terminado==0.11.1
testpath==0.5.0
threadpoolctl==2.2.0
tokenizers==0.10.3
toml==0.10.2
toolz==0.11.1
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.9.0
torchvision==0.10.0
torchaudio==0.9.0
tornado==6.1
tqdm==4.62.2
traitlets==5.1.0
transformers==4.10.2
typed-ast==1.4.3
typing-extensions==3.10.0.2
tzlocal==3.0
urllib3==1.26.6
validators==0.18.2
watchdog==2.1.5
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
wrapt==1.12.1
zipp==3.5.0
